+++ ChangeLog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 - Updated composer.json (01/09/2018)
+- Updated `README.md` (01/09/2018)
+- Added `UPGRADE.md` (01/09/2018)
 
 v1.16.2
 -------